+++ src/main/java/pro/sky/finalprojectsky/repository/ImageRepository.java
@@ -6,7 +6,7 @@ import pro.sky.finalprojectsky.entity.Image;
 /**
  * Repository ImageRepository (advertisement image/изображение в объявлениях).
  */
-public interface ImageRepository extends JpaRepository<Image, Long> {
-    Image findByAdsId(long adsId);
+public interface ImageRepository extends JpaRepository<Image, Integer> {
+    Image findByAdsId(Integer adsId);
 
 }
